LayoutTests/platform/glib/http/tests/websocket/tests/hybi/closed-port-delay-expected.txt

CONSOLE MESSAGE: The server did not accept the WebSocket handshake.
CONSOLE MESSAGE: The server did not accept the WebSocket handshake.
CONSOLE MESSAGE: The server did not accept the WebSocket handshake.
Tests that WebSocket close report is delayed to mitigate port scanning attacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Successfully delayed reporting of closed WebSocket connection
PASS successfullyParsed is true

TEST COMPLETE
